+++ settings.py
@@ -1,0 +1,18 @@
+# Game configuration constants
+
+# Screen dimensions and framerate
+WIDTH, HEIGHT = 800, 600
+FPS = 60
+
+# Player and bullet behaviour
+PLAYER_SPEED = 5
+BULLET_SPEED = -8
+ALIEN_BULLET_SPEED = 4
+
+# Alien movement tuning
+INITIAL_ALIEN_SPEED = 1
+ALIEN_DROP_INCREASE = 0.5
+
+# Scoring rules
+POINTS_PER_ALIEN = 10
+SPEED_UP_SCORE = 100
